+++ centralFederatedClient/src/runCoordinator/eventHandlers/runStart/provisionRun/provisionRun.ts
@@ -72,6 +72,7 @@ export async function provisionRun({
       portBindings: [],
       commandsToRun: ['python', '/workspace/system/entry_provision.py'],
       onContainerExitSuccess: async (containerId) => resolve(undefined),
+      onContainerExitError: async (_, error) => reject(new Error(error)),
     })
   })