+++ desktopApp/electronApp/CHANGELOG.md
@@ -1,5 +1,27 @@
 # neuroflame
 
+## 0.8.0
+
+### Minor Changes
+
+- vault access
+
+### Patch Changes
+
+- Updated dependencies
+  - edge-federated-client@1.7.0
+
+## 0.7.0
+
+### Minor Changes
+
+- vault update
+
+### Patch Changes
+
+- Updated dependencies
+  - edge-federated-client@1.6.0
+
 ## 0.6.0
 
 ### Minor Changes